requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-gnupg==0.3.7
Werkzeug==0.10.4
wheel==0.24.0
